+++ effectful/CHANGELOG.md
@@ -17,6 +17,8 @@
   experience.
 * Properly roll back changes made to the environment when `OnEmptyRollback`
   policy for the `NonDet` effect is selected.
+* Add a `SeqForkUnlift` strategy to support running unlifting functions outside
+  of the scope of effects they capture.
 * **Breaking changes**:
   - `localSeqLend`, `localLend`, `localSeqBorrow` and `localBorrow` now take a
     list of effects instead of a single one.